Revert useDarkMode hook back to useGlobalContext
