+++ src/route/Headers.tsx
@@ -2,10 +2,10 @@ import React, { useState, useContext } from "react";
 import { map } from "lodash-es";
 
 import Label from "../common/Label";
-import { useDarkMode } from "../hooks/use-dark-mode";
+import { useGlobalContext } from "../context/GlobalContext";
 
 export default function Headers({ headers }: { headers: Record<string, any> }) {
-  const [darkMode] = useDarkMode();
+  const { darkMode } = useGlobalContext();
   const [collapsed, setCollapsed] = useState(true);
   if (!headers) {
     return null;
